requirements/_base.txt
-r _base.txt

django-debug-toolbar==3.2
docutils==0.16
pytest==6.2.2
pytest-django==4.1.0
## The following requirements were added by pip freeze:
asgiref==3.3.1
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
Django==3.1.6
django-cors-headers==3.7.0
django-money==1.3.1
djangorestframework==3.12.2
idna==2.10
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
py-moneyed==0.8.0
pyparsing==2.4.7
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.26.3
